+++ R/mappeR-package.R
@@ -2,12 +2,6 @@
 "_PACKAGE"
 
 ## usethis namespace: start
-#' @importFrom grDevices colorRampPalette
-#' @importFrom grDevices rgb
-#' @importFrom igraph delete_edges
-#' @importFrom igraph E
-#' @importFrom igraph graph_from_data_frame
-#' @importFrom igraph set_vertex_attr
 #' @importFrom stats as.dist
 #' @importFrom stats cophenetic
 #' @importFrom stats cutree